+++ packages/instantsearch.js/test/createInstantSearch.ts
@@ -2,7 +2,7 @@ import { createSearchClient } from '@instantsearch/mocks';
 import algoliasearchHelper from 'algoliasearch-helper';
 
 import { INSTANTSEARCH_FUTURE_DEFAULTS } from '../src/lib/InstantSearch';
-import { defer } from '../src/lib/utils';
+import { defer, now } from '../src/lib/utils';
 import { index } from '../src/widgets';
 
 import type { InstantSearch } from '../src/types';
@@ -44,6 +44,7 @@ export const createInstantSearch = (
     _stalledSearchDelay: 200,
     _searchStalledTimer: null,
     _initialUiState: {},
+    _initialOptions: { indexName, searchClient: client },
     _initialResults: null,
     _initialChatStates: null,
     _createURL: jest.fn(() => '#'),
@@ -78,6 +79,7 @@ export const createInstantSearch = (
     emit: jest.fn(),
     listenerCount: jest.fn(),
     sendEventToInsights: jest.fn(),
+    _createdAt: now(),
     future: {
       ...INSTANTSEARCH_FUTURE_DEFAULTS,
       ...(args.future || {}),