+++ packages/instantsearch.css/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.18.0](https://github.com/algolia/instantsearch/compare/instantsearch.css@8.17.0...instantsearch.css@8.18.0) (2026-06-19)
+
+
+### Features
+
+* **chat:** add parseMarkdown opt-out to render message text as plain… ([#7073](https://github.com/algolia/instantsearch/issues/7073)) ([d9add3a](https://github.com/algolia/instantsearch/commit/d9add3a8000e3163f3206fce5e077766f378fcb3))
+
+
+
+
+
+# [8.17.0](https://github.com/algolia/instantsearch/compare/instantsearch.css@8.16.2...instantsearch.css@8.17.0) (2026-06-02)
+
+
+### Features
+
+* **chat:** split chat trigger widget ([#6966](https://github.com/algolia/instantsearch/issues/6966)) ([379084c](https://github.com/algolia/instantsearch/commit/379084cd459a0d8b0f4f2a55b0dae336cf331008))
+
+
+
+
+
 ## [8.16.2](https://github.com/algolia/instantsearch/compare/instantsearch.css@8.16.1...instantsearch.css@8.16.2) (2026-05-26)