+++ packages/vue-instantsearch/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.27.0](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.11...vue-instantsearch@4.27.0) (2026-07-07)
+
+
+### Bug Fixes
+
+* **pagination:** mark the current page with aria-current ([#7103](https://github.com/algolia/instantsearch/issues/7103)) ([f3db948](https://github.com/algolia/instantsearch/commit/f3db948b884e64c1b121fbb3801b10959ddb07ea)), closes [#7098](https://github.com/algolia/instantsearch/issues/7098)
+* **refinement-list:** expose show more/less state via aria-expanded ([#7102](https://github.com/algolia/instantsearch/issues/7102)) ([7d82c2f](https://github.com/algolia/instantsearch/commit/7d82c2f101b4a69b3d613d9cfb9be44e0881a4ea)), closes [#7098](https://github.com/algolia/instantsearch/issues/7098)
+* **searchbox:** give submit and reset buttons a reliable accessible name ([#7100](https://github.com/algolia/instantsearch/issues/7100)) ([59aba4f](https://github.com/algolia/instantsearch/commit/59aba4f7cf422e814ac3df35fbe8eada434ef30a)), closes [#7098](https://github.com/algolia/instantsearch/issues/7098) [#7098](https://github.com/algolia/instantsearch/issues/7098)
+
+
+### Features
+
+* **refinement-list:** allow a custom accessible name for the show more button ([#7101](https://github.com/algolia/instantsearch/issues/7101)) ([cf04cb7](https://github.com/algolia/instantsearch/commit/cf04cb758b7e2ba2c8e0c07c9ec8157491bcaee5)), closes [#7098](https://github.com/algolia/instantsearch/issues/7098)
+* **stats:** announce result changes to screen readers ([#7099](https://github.com/algolia/instantsearch/issues/7099)) ([0821eb8](https://github.com/algolia/instantsearch/commit/0821eb846b9bbf6ec3740b39050798eaec133855)), closes [#7098](https://github.com/algolia/instantsearch/issues/7098)
+
+
+
+
+
+## [4.26.11](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.9...vue-instantsearch@4.26.11) (2026-06-30)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
+## [4.26.10](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.9...vue-instantsearch@4.26.10) (2026-06-25)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
+## [4.26.9](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.8...vue-instantsearch@4.26.9) (2026-06-19)
+
+
+### Bug Fixes
+
+* **range-slider:** make handles keyboard-movable without snap points ([#7071](https://github.com/algolia/instantsearch/issues/7071)) ([55bd627](https://github.com/algolia/instantsearch/commit/55bd627dbacabd538b92aa1c8654ec3631e59ece))
+
+
+
+
+
+## [4.26.8](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.7...vue-instantsearch@4.26.8) (2026-06-09)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
+## [4.26.7](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.6...vue-instantsearch@4.26.7) (2026-06-03)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
+## [4.26.6](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.5...vue-instantsearch@4.26.6) (2026-06-02)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
 ## [4.26.5](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.4...vue-instantsearch@4.26.5) (2026-05-26)
 
 **Note:** Version bump only for package vue-instantsearch